CA-382045: VM guidance is missed when merging un-applied guidances
After applying updates, the pending guidances from the update was
merged with any un-applied guidances, and then were set into DB, and
the VM guidance: RestartDeviceModel was missed.

However it is not easy to "merge" un-applied VM guidance with guidances
from update, as these guidances are for each individule VM, with a
single "RestartDeviceModel", you can not tell which VMs needs this
guidance, as some VM might have already applied this guidance.

So in this change, instead of trying to merge the guidances, just
"update" the guidances from the update one by one:
1. add the guidance to DB
2. remove any guidance which is preceded by this guidance from DB.

It works because the guidances from update are filtered (by
resort_guidances), no preceded guidances will exist together with the
preceding one.

Signed-off-by: Gang Ji <gang.ji@citrix.com>